:recycle: Class Seperation
seperate each and every class into it's own file
