+++ src/Entities/EnemyLaser.js
@@ -1,0 +1,9 @@
+import 'phaser';
+import Entity from './Entities';
+
+export default class EnemyLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprLaserEnemy0');
+    this.body.velocity.y = 200;
+  }
+}
